+++ example/src/main/java/arez/dom/example/DocumentVisibilityExample.java
@@ -1,9 +1,9 @@
 package arez.dom.example;
 
+import akasha.Console;
 import arez.Arez;
 import arez.dom.DocumentVisibility;
 import com.google.gwt.core.client.EntryPoint;
-import elemental2.dom.DomGlobal;
 
 public class DocumentVisibilityExample
   implements EntryPoint
@@ -11,6 +11,6 @@ public class DocumentVisibilityExample
   public void onModuleLoad()
   {
     final DocumentVisibility v = DocumentVisibility.create();
-    Arez.context().observer( () -> DomGlobal.console.log( "Document Visibility: " + v.getVisibility() ) );
+    Arez.context().observer( () -> Console.log( "Document Visibility: " + v.getVisibility() ) );
   }
 }