+++ example/src/main/java/arez/dom/example/GeoPositionExample.java
@@ -1,10 +1,11 @@
 package arez.dom.example;
 
+import akasha.Global;
+import akasha.HTMLElement;
 import arez.Arez;
 import arez.dom.GeoPosition;
 import arez.dom.Position;
 import com.google.gwt.core.client.EntryPoint;
-import elemental2.dom.DomGlobal;
 import jsinterop.base.Js;
 
 public class GeoPositionExample
@@ -14,19 +15,17 @@ public class GeoPositionExample
   {
     final GeoPosition geoPosition = GeoPosition.create();
 
-    DomGlobal.document.querySelector( "#watch" ).onclick = e -> {
+    ( (HTMLElement) Global.document().querySelector( "#watch" ) ).onclick = e -> {
       Arez.context().observer( () -> {
         final int status = geoPosition.getStatus();
         final String errorMessage = geoPosition.getErrorMessage();
         final Position position = geoPosition.getPosition();
         Js.debugger();
-        DomGlobal.document.querySelector( "#status" ).textContent =
+        Global.document().querySelector( "#status" ).textContent =
           "GeoPosition: " + status +
           ( null != position ? " Position: " + position.getLatitude() + ", " + position.getLongitude() : "" ) +
           ( null != errorMessage ? " ErrorMessage: " + errorMessage : "" );
       } );
-      return null;
     };
-
   }
 }