+++ example/src/main/java/arez/dom/example/IdleStatusExample.java
@@ -1,9 +1,9 @@
 package arez.dom.example;
 
+import akasha.Console;
 import arez.Arez;
 import arez.dom.IdleStatus;
 import com.google.gwt.core.client.EntryPoint;
-import elemental2.dom.DomGlobal;
 
 public class IdleStatusExample
   implements EntryPoint
@@ -12,9 +12,6 @@ public class IdleStatusExample
   public void onModuleLoad()
   {
     final IdleStatus idleStatus = IdleStatus.create();
-    Arez.context().observer( () -> {
-      final String message = "Interaction Status: " + ( idleStatus.isIdle() ? "Idle" : "Active" );
-      DomGlobal.console.log( message );
-    } );
+    Arez.context().observer( () -> Console.log( "Interaction Status: " + ( idleStatus.isIdle() ? "Idle" : "Active" ) ) );
   }
 }