+++ example/src/main/java/arez/dom/example/NetworkStatusExample.java
@@ -3,7 +3,7 @@ package arez.dom.example;
 import arez.Arez;
 import arez.dom.NetworkStatus;
 import com.google.gwt.core.client.EntryPoint;
-import elemental2.dom.DomGlobal;
+import akasha.Global;
 
 public class NetworkStatusExample
   implements EntryPoint
@@ -13,7 +13,7 @@ public class NetworkStatusExample
   {
     final NetworkStatus networkStatus = NetworkStatus.create();
     Arez.context().observer( () ->
-                               DomGlobal.document.querySelector( "#network" ).textContent =
+                               Global.document().querySelector( "#network" ).textContent =
                                  "Network Status: " + ( networkStatus.isOnLine() ? "Online" : "Offline" ) );
   }
 }