+++ src/main/java/arez/dom/BrowserLocation.java
@@ -1,5 +1,9 @@
 package arez.dom;
 
+import akasha.Global;
+import akasha.HashChangeEvent;
+import akasha.HashChangeEventListener;
+import akasha.Location;
 import arez.ComputableValue;
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
@@ -10,9 +14,6 @@ import arez.annotations.Memoize;
 import arez.annotations.Observable;
 import arez.annotations.OnActivate;
 import arez.annotations.OnDeactivate;
-import elemental2.dom.DomGlobal;
-import elemental2.dom.Event;
-import elemental2.dom.EventListener;
 import java.util.Objects;
 import javax.annotation.Nonnull;
 
@@ -34,7 +35,7 @@ import javax.annotation.Nonnull;
 public abstract class BrowserLocation
 {
   @Nonnull
-  private final EventListener _listener = this::onHashChangeEvent;
+  private final HashChangeEventListener _listener = this::onHashChangeEvent;
   /**
    * The location according to the application.
    */
@@ -146,13 +147,13 @@ public abstract class BrowserLocation
   @OnActivate
   void onBrowserLocationActivate()
   {
-    DomGlobal.window.addEventListener( "hashchange", _listener, false );
+    Global.addHashchangeListener( _listener, false );
   }
 
   @OnDeactivate
   void onBrowserLocationDeactivate()
   {
-    DomGlobal.window.removeEventListener( "hashchange", _listener, false );
+    Global.removeHashchangeListener( _listener, false );
   }
 
   @ComputableValueRef
@@ -169,7 +170,7 @@ public abstract class BrowserLocation
     }
   }
 
-  private void onHashChangeEvent( @Nonnull final Event e )
+  private void onHashChangeEvent( @Nonnull final HashChangeEvent e )
   {
     if ( _preventDefault )
     {
@@ -181,24 +182,23 @@ public abstract class BrowserLocation
   @Nonnull
   private String getHash()
   {
-    final String hash = DomGlobal.window.location.hash;
-    return null == hash ? "" : hash.substring( 1 );
+    return Global.location().hash.substring( 1 );
   }
 
   private void setHash( @Nonnull final String hash )
   {
+    final Location location = Global.location();
     if ( 0 == hash.length() )
     {
       /*
        * This code is needed to remove the stray #.
        * See https://stackoverflow.com/questions/1397329/how-to-remove-the-hash-from-window-location-url-with-javascript-without-page-r/5298684#5298684
        */
-      final String url = DomGlobal.window.location.pathname + DomGlobal.window.location.search;
-      DomGlobal.window.history.pushState( "", DomGlobal.document.title, url );
+      Global.history().pushState( "", Global.document().title, location.pathname + location.search );
     }
     else
     {
-      DomGlobal.window.location.hash = hash;
+      location.hash = hash;
     }
   }
 }